\name{ki}
\alias{ki}
\docType{data}
\title{Konjunkturinstituitet's forecasts}
\description{Forecasts from Konjunkturinstitutet of the inflation measured by CPI and KPIX with forecast origins 2001:Q1-2005:Q2. At each of the 18 forecast origin, point forecasts 1-21 months ahead as well as outcome data.}
\usage{ki}
\format{\code{ki} consists of two lists \code{cpi} and \code{kpix}, each of which is a list of 25 time series objects, one for each forecast origin 1999:Q1-2005:Q2. Each of the time series objects consists of 2 variables; point forecast and outcome.}
\author{Michael Lundholm}
\keyword{datasets}
