\name{kpix.outcome}
\alias{kpix.outcome}
\docType{data}
\title{Swedish KPIX inflation 1999:M05-2007:M07}
\description{Annual changes in the Swedish inflation measured by CPI 1999:M05-2007:M07.}
\usage{kpix.outcome}
\format{A time series object with frequency 12 containing 99 observations with starting date 1999:M05.}
\author{Michael Lundholm}
\keyword{datasets}
