\name{sifds}
\alias{sifds}
\alias{ki.error}
\alias{rb.error}
\docType{package}
\title{Swedish inflation forecast data set}
\description{A package containing data sets consisting of 25
forecasts (1999-2005) by Sveriges Riksbank
(RB) and 18 forecasts (2001-2005) by
Konjunkturinstitutet (KI) for Swedish
inflation rates measured as CPI and KPIX.
Lundholm (2010) contains a detailed description
sources and how the data set is constructed.}
\usage{data(sifds)}
\author{Michael Lundholm}
\references{Lundholm, M., (2010), \pkg{sifds}: Swedish inflation forecast data set 1999:Q2-2005:Q2, Research papers in economics 2010:4, Department of Economics, Stockholm University (http://swopec.hhs.se).}
\keyword{ sifds }
